Creating flask backend api
Testing out a get request from flask. Running the rest script will call functions to scrape reddit. This then returned to the app calling the api in a json file. After this full works and i clean up the code i will  then connect this to my react front end. The aim then is that the the user can use the search field and look up a  subreddit, the posts will be displayed in the UI.