Working on API get and post requests
Trying to render the JSON data from my flask api in react but was having trouble getting that working. Trying now to do a post request , the user enters a subreddits name into the search bar, sending a  get request to the server which should then return the relevant data.